+++ src/Entity/Notification/RefundNotifying.php
@@ -10,6 +10,7 @@ use MercadoPago\PP\Sdk\Common\AbstractEntity;
  * @property int $id
  * @property bool $notifying
  * @property float $amount
+ * @property float $original_currency_amount
 
  * @package MercadoPago\PP\Sdk\Entity\Notification
  */
@@ -29,4 +30,9 @@ class RefundNotifying extends AbstractEntity
      * @var float
      */
     protected $amount;
+
+    /**
+     * @var float
+     */
+    protected $original_currency_amount;
 }