+++ tests/unit/Mock/ExchangeMock.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace MercadoPago\PP\Sdk\Tests\Unit\Mock;
+
+/**
+ * Class ExchangeMock
+ */
+class ExchangeMock
+{
+    const EXCHANGE_MOCK = [
+        "currency_base" => "USD",
+        "currency_quote" => "BRL",
+        "rate" => 5.3097,
+        "creation_date" => "2024-06-04T14:05:24.000+00:00"
+    ];
+}