+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V2.11.0
+### Added
+- Add new entity and client to get exchange in payment-methods service
+
 ## V2.10.1
 ### Changed
-changes the read method of the payment and preference entities, to pass on the `shouldTheExpectedResponseBeMappedOntoTheEntity` attribute in the call to the read method of the parent class
+- changes the read method of the payment and preference entities, to pass on the `shouldTheExpectedResponseBeMappedOntoTheEntity` attribute in the call to the read method of the parent class
 
 ## V2.10.0
 ### Added